Object-oriented theoreticalDatabase; moving into Comparisons
