+++ PS_0.00/PS_0.00/ProteoformFamily.cs
@@ -8,5 +8,17 @@ namespace PS_0._00
 {
     class ProteoformFamily
     {
+        public int lysine_count;
+
+        
+    }
+
+    class ProteoformCommunity
+    {
+        public ProteoformCommunity(List<ExperimentalProteoform> experimental_pfs, List<TheoreticalProteoform> theoretical_pfs)
+        {
+            //construct full adjacency matrix with the special considerations for ET and EE
+
+        }
     }
 }
